Use case 2 add on

+++ app/P2P/messages/messages.go
@@ -3,14 +3,14 @@ package messages
 // client message
 type P2PClientMessage struct {
 	PeerID    string `json:"peer_id"`
-	MessageID int    `json:"message_id"`
+	MessageID uint64 `json:"message_id"`
 	Method    string `json:"method"`
-	Payload   []byte `json:"payload"`
+	Payload   []int  `json:"payload"`
 }
 
 // server message
 type P2PServerMessasge struct {
 	MessageID int    `json:"message_id"`
 	Method    string `json:"method"`
-	Payload   []byte `json:"payload"`
+	Payload   []int  `json:"payload"`
 }
